Added some logging changes,  moved support classes form all.app to omar.services
