+++ apps/omar-services-app/grails-app/conf/logback.groovy
@@ -16,6 +16,7 @@ appender('STDOUT', ConsoleAppender) {
 // }
 
 root(ERROR, ['STDOUT']) //, 'FILE'])
+logger('omar', INFO, ['STDOUT'], false) 
 
 def targetDir = BuildSettings.TARGET_DIR
 if (Environment.isDevelopmentMode() && targetDir) {
